Refactored keyword extraction model to use count file.
